updated end

+++ app/Controllers/import.php
@@ -16,35 +16,40 @@ class import extends controller
                 $file = './Uploads/FileImport/' . $file_name;
                 $spreadsheet = \PhpOffice\PhpSpreadsheet\IOFactory::load($file);
                 $data = $spreadsheet->getActiveSheet()->toArray();
-
-                try {
-                    $count = "0";
-                    foreach ($data as $row) {
-                        if ($count > 0) {
-                            $malop = $row['0'];
-                            $maKhoa = $row['1'];
-                            $heDaoTao = $row['2'];
-                            $nienKhoa = $row['3'];
-                            $isExistClass = $this->Model("lop")->isExistClass($malop, $maKhoa);
-                            if ($isExistClass[0]['number'] == 0) {
-                                $this->Model('lop')->addListClassByImportFile($malop, $maKhoa, $heDaoTao, $nienKhoa);
-                                $_SESSION['status'] = "File import dữ liệu thành công ";
-                                $_SESSION['status_code'] = "success";
-                                header("location: " . _WEB_ROOT_ . "/Admin/addClassPage");
-                            } else {
-                                $_SESSION['status'] = "Một số lớp đã tồn tại trong hệ thống ";
-                                $_SESSION['status_code'] = "error";
-                                header("location: " . _WEB_ROOT_ . "/Admin/addClassPage");
-                            }
-                        } else {
-                            $count = "1";
-                        }
-                    }
-                    unlink('./Uploads/FileImport/' . $file_name);
-                } catch (Exception $ex) {
-                    $_SESSION['status'] = $ex->getMessage() . "Uploads thất bại";
+                if (empty($data[0][0])) {
+                    $_SESSION['status'] = "Không tồn tại dữ liệu trong file.";
                     $_SESSION['status_code'] = "error";
                     header("location: " . _WEB_ROOT_ . "/Admin/addClassPage");
+                } else {
+                    try {
+                        $count = "0";
+                        foreach ($data as $row) {
+                            if ($count > 0) {
+                                $malop = $row['0'];
+                                $maKhoa = $row['1'];
+                                $heDaoTao = $row['2'];
+                                $nienKhoa = $row['3'];
+                                $isExistClass = $this->Model("lop")->isExistClass($malop, $maKhoa);
+                                if ($isExistClass[0]['number'] == 0) {
+                                    $this->Model('lop')->addListClassByImportFile($malop, $maKhoa, $heDaoTao, $nienKhoa);
+                                    $_SESSION['status'] = "File import dữ liệu thành công ";
+                                    $_SESSION['status_code'] = "success";
+                                    header("location: " . _WEB_ROOT_ . "/Admin/addClassPage");
+                                } else {
+                                    $_SESSION['status'] = "Một số lớp đã tồn tại trong hệ thống ";
+                                    $_SESSION['status_code'] = "warning";
+                                    header("location: " . _WEB_ROOT_ . "/Admin/addClassPage");
+                                }
+                            } else {
+                                $count = "1";
+                            }
+                        }
+                    } catch (Exception $ex) {
+                        $_SESSION['status'] = "Import thất bại, Vui lòng kiểm tra lại dữ liệu đầu vào.";
+                        $_SESSION['status_code'] = "error";
+                        header("location: " . _WEB_ROOT_ . "/Admin/addClassPage");
+                    }
+                    unlink('./Uploads/FileImport/' . $file_name);
                 }
             } else {
                 $_SESSION['status'] = "File import không đúng định dạng vui lòng kiểm tra lại";
